+++ src/main/scala/satify/update/Update.scala
@@ -1,5 +1,6 @@
 package satify.update
 
+import satify.dsl.Reflection.*
 import satify.model.CNF.Symbol
 import satify.model.{Expression, State, Variable}
 import satify.update.Message.*
@@ -14,10 +15,10 @@ object Update:
     message match
       case Input(char) => model
       case Solve(input) =>
-        val exp = reflect(processInput(input))
+        val exp = reflect(input)
         State()
       case Convert(input) =>
-        val exp = reflect(processInput(input))
+        val exp = reflect(input)
         val cnf = tseitin(exp)
         State(exp, cnf)
       case Import(file) =>
@@ -26,21 +27,3 @@ object Update:
         s.close()
         val optCnf = parse(lines)
         State(Expression.Symbol("NO EXP"), optCnf.getOrElse(Symbol(Variable("NO CNF FOUND"))))
-
-  private def processInput(input: String): String =
-    // TODO: link these operators to the ones in the DSL
-    val operators = List("and", "or", "not", "=>", "\\/", "/\\", "(", ")", "^")
-    input.trim
-      .split("[ ()]")
-      .filterNot(_.isBlank)
-      .map(symbol => if !operators.contains(symbol) then s"\"$symbol\"" else symbol)
-      .reduce((a, b) => s"$a $b")
-
-  private def reflect(code: String): Expression =
-    val imports =
-      """import satify.model.{Expression}
-        |import satify.dsl.DSL.{*, given}
-        |""".stripMargin
-    println(imports + code)
-    // TODO: does not work with only one symbol
-    dotty.tools.repl.ScriptEngine().eval(imports + code).asInstanceOf[Expression]
